Restore original story scene styling
